modified:   index.html 	modified:   photography.html 	modified:   scripts.js 	modified:   styles.css
I swear, this static website has taken more time than a lot of my analytics projects. I'm not sure if that's a good thing or a bad thing. I'm going to go with good thing because I'm learning a lot about web development. I'm also learning that I'm not a fan of web development.